Allow setting listing price

+++ app/graphql/inputs/offer_input.rb
@@ -1,4 +1,4 @@
 class Inputs::OfferInput < Types::BaseInputObject
-  argument :price, Integer, required: true
+  argument :price, Float, required: true
   argument :listing_id, Integer, required: true
 end
